Init refactor folders and files
