+++ coursemapper-kg/Pipfile
@@ -30,7 +30,6 @@ tqdm = "*"
 requests = "*"
 scipy = "*"
 protobuf = "*"
-install = "*"
 markupsafe = "==1.1.1"
 flair = "*"
 spacy = "*"
