+++ webserver/src/activity-logger/generator/annotation-generator.js
