+++ webserver/src/activity-logger/generator/auth-generator.js
