+++ webserver/src/activity-logger/generator/channel-generator.js
