+++ webserver/src/activity-logger/generator/course-generator.js
