+++ webserver/src/activity-logger/generator/material-generator.js
