+++ webserver/src/activity-logger/generator/topic-generator.js
