+++ webserver/src/activity-logger/logger/channel-logger.js
@@ -1,15 +1,15 @@
-const statementFactory = require("../statementsFactory/topic.statementsFactory");
+const statementFactory = require("../generator/channel-generator");
 const lrs = require("../lrs/lrs");
-const controller = require("../controller.xAPILogger");
+const controller = require("../controller/activity-controller");
 const ORIGIN = process.env.ORIGIN;
 const notifications = require("../../middlewares/Notifications/notifications");
 
-export const newTopic = async (req, res, next) => {
+export const newChannel = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  const statement = statementFactory.getTopicCreationStatement(
+  const statement = statementFactory.getChannelCreationStatement(
     req.locals.user,
-    req.locals.topic,
-    origin
+    req.locals.channel,
+    origin,
   );
 
   const notificationInfo = notifications.generateNotificationInfo(req);
@@ -18,32 +18,32 @@ export const newTopic = async (req, res, next) => {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
+    //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
   } catch (err) {
     res.status(500).send({ error: "Error saving statement to mongo", err });
   }
-
   next();
 };
 
-export const deleteTopic = async (req, res, next) => {
+export const deleteChannel = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  const statement = statementFactory.getTopicDeletionStatement(
+  const statement = statementFactory.getChannelDeletionStatement(
     req.locals.user,
-    req.locals.topic,
-    origin
+    req.locals.channel,
+    origin,
   );
   const notificationInfo = notifications.generateNotificationInfo(req);
   const sent = await lrs.sendStatementToLrs(statement);
-
   try {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
+    //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
   } catch (err) {
     res.status(500).send({ error: "Error saving statement to mongo", err });
@@ -51,38 +51,39 @@ export const deleteTopic = async (req, res, next) => {
   next();
 };
 
-export const getTopic = async (req, res) => {
+export const editChannel = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  const statement = statementFactory.getTopicAccessStatement(
+  const statement = statementFactory.getChannelEditStatement(
     req.locals.user,
-    req.locals.topic,
-    origin
+    req.locals.newChannel,
+    req.locals.oldChannel,
+    origin,
   );
+  const notificationInfo = notifications.generateNotificationInfo(req);
+  const sent = await lrs.sendStatementToLrs(statement);
+  try {
+    const activity = await controller.saveStatementToMongo(
+      statement,
+      sent,
+      notificationInfo,
+    );
+    //Add activity to req.locals so it can be used in the notification
+    req.locals.activity = activity;
+  } catch (err) {
+    res.status(500).send({ error: "Error saving statement to mongo", err });
+  }
+  next();
+};
+
+export const getChannel = async (req, res) => {
+  const origin = req.get("origin") ? req.get("origin") : ORIGIN;
+  const statement = statementFactory.getChannelAccessStatement(
+    req.locals.user,
+    req.locals.channel,
+    origin,
+  );
+
   const sent = await lrs.sendStatementToLrs(statement);
   controller.saveStatementToMongo(statement, sent);
   res.status(200).send(req.locals.response);
-};
-
-export const editTopic = async (req, res, next) => {
-  const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  const statement = statementFactory.getTopicEditStatement(
-    req.locals.user,
-    req.locals.newTopic,
-    req.locals.oldTopic,
-    origin
-  );
-
-  const notificationInfo = notifications.generateNotificationInfo(req);
-  const sent = await lrs.sendStatementToLrs(statement);
-  try {
-    const activity = await controller.saveStatementToMongo(
-      statement,
-      sent,
-      notificationInfo
-    );
-    req.locals.activity = activity;
-  } catch (err) {
-    res.status(500).send({ error: "Error saving statement to mongo", err });
-  }
-  next();
 };
