+++ webserver/src/activity-logger/logger/reply-logger.js
@@ -1,25 +1,25 @@
-const statementFactory = require("../statementsFactory/annotation.statementsFactory");
+const statementFactory = require("../generator/reply-generator");
 const lrs = require("../lrs/lrs");
-const controller = require("../controller.xAPILogger");
+const controller = require("../controller/activity-controller");
 const ORIGIN = process.env.ORIGIN;
 const notifications = require("../../middlewares/Notifications/notifications");
 
-export const newAnnotation = async (req, res, next) => {
+export const newReply = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
   let statement;
   if (!req.locals.annotation.tool) {
-    statement = statementFactory.getCommentCreationStatement(
+    statement = statementFactory.getReplyToCommentCreationStatement(
       req.locals.user,
       req.locals.annotation,
-      req.locals.material,
-      origin
+      req.locals.reply,
+      origin,
     );
   } else {
-    statement = statementFactory.getAnnotationCreationStatement(
+    statement = statementFactory.getReplyToAnnotationCreationStatement(
       req.locals.user,
       req.locals.annotation,
-      req.locals.material,
-      origin
+      req.locals.reply,
+      origin,
     );
   }
   const notificationInfo = notifications.generateNotificationInfo(req);
@@ -28,7 +28,7 @@ export const newAnnotation = async (req, res, next) => {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
     //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
@@ -38,29 +38,21 @@ export const newAnnotation = async (req, res, next) => {
   next();
 };
 
-export const deleteAnnotation = async (req, res, next) => {
+export const deleteReply = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  let statement;
-  if (!req.locals.annotation.tool) {
-    statement = statementFactory.getCommentDeletionStatement(
-      req.locals.user,
-      req.locals.annotation,
-      origin
-    );
-  } else {
-    statement = statementFactory.getAnnotaionDeletionStatement(
-      req.locals.user,
-      req.locals.annotation,
-      origin
-    );
-  }
+  const statement = statementFactory.getReplyDeletionStatement(
+    req.locals.user,
+    req.locals.reply,
+    origin,
+    req.locals.annotation,
+  );
   const notificationInfo = notifications.generateNotificationInfo(req);
   const sent = await lrs.sendStatementToLrs(statement);
   try {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
     //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
@@ -70,37 +62,23 @@ export const deleteAnnotation = async (req, res, next) => {
   next();
 };
 
-export const likeAnnotation = async (req, res, next) => {
+export const likeReply = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
   let statement;
   if (req.locals.like) {
-    if (!req.locals.annotation.tool) {
-      statement = statementFactory.getCommentLikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    } else {
-      statement = statementFactory.getAnnotationLikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    }
+    statement = statementFactory.getReplyLikeStatement(
+      req.locals.user,
+      req.locals.reply,
+      origin,
+      req.locals.annotation,
+    );
   } else {
-    if (!req.locals.annotation.tool) {
-      statement = statementFactory.getCommentUnlikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    } else {
-      statement = statementFactory.getAnnotationUnlikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    }
+    statement = statementFactory.getReplyUnlikeStatement(
+      req.locals.user,
+      req.locals.reply,
+      origin,
+      req.locals.annotation,
+    );
   }
   const notificationInfo = notifications.generateNotificationInfo(req);
   const sent = await lrs.sendStatementToLrs(statement);
@@ -108,7 +86,7 @@ export const likeAnnotation = async (req, res, next) => {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
     //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
@@ -118,37 +96,23 @@ export const likeAnnotation = async (req, res, next) => {
   next();
 };
 
-export const dislikeAnnotation = async (req, res, next) => {
+export const dislikeReply = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
   let statement;
   if (req.locals.dislike) {
-    if (!req.locals.annotation.tool) {
-      statement = statementFactory.getCommentDislikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    } else {
-      statement = statementFactory.getAnnotationDislikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    }
+    statement = statementFactory.getReplyDislikeStatement(
+      req.locals.user,
+      req.locals.reply,
+      origin,
+      req.locals.annotation,
+    );
   } else {
-    if (!req.locals.annotation.tool) {
-      statement = statementFactory.getCommentUndislikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    } else {
-      statement = statementFactory.getAnnotationUndislikeStatement(
-        req.locals.user,
-        req.locals.annotation,
-        origin
-      );
-    }
+    statement = statementFactory.getReplyUndislikeStatement(
+      req.locals.user,
+      req.locals.reply,
+      origin,
+      req.locals.annotation,
+    );
   }
   const notificationInfo = notifications.generateNotificationInfo(req);
   const sent = await lrs.sendStatementToLrs(statement);
@@ -156,7 +120,7 @@ export const dislikeAnnotation = async (req, res, next) => {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
     //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
@@ -166,31 +130,22 @@ export const dislikeAnnotation = async (req, res, next) => {
   next();
 };
 
-export const editAnnotation = async (req, res, next) => {
+export const editReply = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  let statement;
-  if (!req.locals.oldAnnotation.tool) {
-    statement = statementFactory.getCommentEditStatement(
-      req.locals.user,
-      req.locals.newAnnotation,
-      req.locals.oldAnnotation,
-      origin
-    );
-  } else {
-    statement = statementFactory.getAnnotationEditStatement(
-      req.locals.user,
-      req.locals.newAnnotation,
-      req.locals.oldAnnotation,
-      origin
-    );
-  }
+  let statement = statementFactory.getReplyEditStatement(
+    req.locals.user,
+    req.locals.newReply,
+    origin,
+    req.locals.annotation,
+  );
+  req.locals.category = "mentionedandreplied";
   const notificationInfo = notifications.generateNotificationInfo(req);
   const sent = await lrs.sendStatementToLrs(statement);
   try {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
     //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
@@ -202,11 +157,11 @@ export const editAnnotation = async (req, res, next) => {
 
 export const newMention = async (req, res, next) => {
   const origin = req.get("origin") ? req.get("origin") : ORIGIN;
-  req.locals.category = "mentionedandreplied";
   const statement = statementFactory.getNewMentionCreationStatement(
     req.locals.user,
+    req.locals.reply,
+    origin,
     req.locals.annotation,
-    origin
   );
   const notificationInfo = notifications.generateNotificationInfo(req);
   const sent = await lrs.sendStatementToLrs(statement);
@@ -214,7 +169,7 @@ export const newMention = async (req, res, next) => {
     const activity = await controller.saveStatementToMongo(
       statement,
       sent,
-      notificationInfo
+      notificationInfo,
     );
     //Add activity to req.locals so it can be used in the notification
     req.locals.activity = activity;
