+++ webserver/src/activity-logger/lrs/lrs.js
@@ -1,4 +1,4 @@
-const { lrs } = require("./lrs.config");
+const { lrs } = require("./lrs-config");
 const SEND_STATEMENT_IN_REALTIME =
   process.env.SEND_STATEMENT_IN_REALTIME === "true";
 
@@ -8,11 +8,11 @@ export const sendStatementToLrs = async (statement) => {
     try {
       response = await lrs.put(
         `/statements?statementId=${statement.id}`,
-        statement
+        statement,
       );
       if (response.status === 204) {
         console.log(
-          `sendStatementToLrs: statement ${statement.id} saved successfully to LRS`
+          `sendStatementToLrs: statement ${statement.id} saved successfully to LRS`,
         );
         return true;
       }
@@ -37,7 +37,7 @@ export const sendStatementsToLrs = async (statements) => {
     response = await lrs.post(`/statements`, statements);
     if (response.status === 200) {
       console.log(
-        `sendStatementsToLrs: ${response.data.length} statemens are saved successfully to LRS`
+        `sendStatementsToLrs: ${response.data.length} statemens are saved successfully to LRS`,
       );
       return response.data;
     }
