+++ webserver/src/routes/channel.routes.js
@@ -1,6 +1,6 @@
 const { authJwt, notifications } = require("../middlewares");
 const controller = require("../controllers/channel.controller");
-const logger = require("../xAPILogger/logger/channel.logger");
+const logger = require("../activity-logger/logger/channel-logger");
 
 module.exports = function (app) {
   app.use(function (req, res, next) {
@@ -14,7 +14,7 @@ module.exports = function (app) {
     "/api/courses/:courseId/channels/:channelId",
     [authJwt.verifyToken, authJwt.isEnrolled],
     controller.getChannel,
-    logger.getChannel
+    logger.getChannel,
   );
 
   // Create a new channel
@@ -26,7 +26,7 @@ module.exports = function (app) {
     logger.newChannel,
     notifications.updateBlockingNotificationsNewChannel,
     notifications.channelCourseUpdateNotificationUsers,
-    notifications.populateUserNotification
+    notifications.populateUserNotification,
   );
 
   // Delete a channel
@@ -37,7 +37,7 @@ module.exports = function (app) {
     controller.deleteChannel,
     logger.deleteChannel,
     notifications.channelCourseUpdateNotificationUsers,
-    notifications.populateUserNotification
+    notifications.populateUserNotification,
   );
 
   // Edit a channel
@@ -48,39 +48,36 @@ module.exports = function (app) {
     controller.editChannel,
     logger.editChannel,
     notifications.channelCourseUpdateNotificationUsers,
-    notifications.populateUserNotification
+    notifications.populateUserNotification,
   );
-
 
   app.post(
     "/api/courses/:courseId/channels/:channelId/indicator",
     [authJwt.verifyToken, authJwt.isModerator],
-    controller.newIndicator
+    controller.newIndicator,
   );
 
   app.delete(
     "/api/courses/:courseId/channels/:channelId/indicator/:indicatorId",
     [authJwt.verifyToken, authJwt.isModerator],
-    controller.deleteIndicator
+    controller.deleteIndicator,
   );
 
   app.get(
     "/api/courses/:courseId/channels/:channelId/indicator",
     [authJwt.verifyToken, authJwt.isEnrolled],
-    controller.getIndicators
+    controller.getIndicators,
   );
 
   app.put(
     "/api/courses/:courseId/channels/:channelId/indicator/:indicatorId/resize/:width/:height",
     [authJwt.verifyToken, authJwt.isModerator],
-    controller.resizeIndicator
+    controller.resizeIndicator,
   );
 
   app.put(
     "/api/courses/:courseId/channels/:channelId/reorder/:newIndex/:oldIndex",
     [authJwt.verifyToken, authJwt.isModerator],
-    controller.reorderIndicators
+    controller.reorderIndicators,
   );
-
-
 };
